% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agencies}
\alias{agencies}
\title{Agencies from State and Federal lists}
\format{
A data frame with 7785 rows and 8 variables
}
\source{
\url{input URL}
}
\usage{
agencies
}
\description{
A dataset containing the names of federal agencies, state agencies, and in some cases, their abbreviations
}
\keyword{datasets}
