% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_compile.R
\name{citation_compile}
\alias{citation_compile}
\title{Compiling JSONs from the Anystyle extraction}
\usage{
citation_compile(ref_dir)
}
\arguments{
\item{x}{file(s) with .JSON extension}
}
\value{
data frame
}
\description{
Reads in JSON files of extracted citations and combines them
}
