% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_extract.R
\name{citation_extract}
\alias{citation_extract}
\title{Anystyle.io extraction}
\usage{
citation_extract(doc_dir, ref_dir, layout)
}
\arguments{
\item{x}{file with .pdf extension}
}
\value{
JSON
}
\description{
Runs PDFs through the \href{https://anystyle.io/}{Anystyle API} and create JSON files for each PDF with the identified citations.
}
\examples{
citation_extraction(doc_dir = 'documents_trial1', ref_dir = 'reference_extracts_trial1_nolayout', layout = "none")

}
