% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{error_references}
\alias{error_references}
\title{A sample of candidate references extracted from a PDF file THAT DOES NOT CURRENTLY WORK IN reference_clean()}
\format{
A data.table object with 1001 rows
}
\description{
A dataset containing tagged references from a PDF extracted using anystyle.io.
This set throws errors when fed into reference_clean(), and is primarily used for future debugging
}
