% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{null2NA}
\alias{null2NA}
\title{Take a list with NULL entries and replace with NA value placeholders}
\usage{
null2NA(l)
}
\description{
Take a list with NULL entries and replace with NA value placeholders
}
