% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_clean.R
\name{reference_clean}
\alias{reference_clean}
\title{Clean Anystyle output}
\usage{
reference_clean(dt)
}
\arguments{
\item{dt}{data table from the reference_compile() function}
}
\value{
data table
}
\description{
Runs through columns of the Anystyle output (c("date", "url", "title", "container", "publisher", "doi")) and cleans them. Steps include identifying the lengths of different lists in each reference to unlist and unnest them to create sensible references.
}
\examples{
cleaned_dt <- reference_clean(dt)

}
