% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_extract.R
\name{reference_extract}
\alias{reference_extract}
\title{Anystyle.io extraction}
\usage{
reference_extract(
  doc_dir = NULL,
  files = NULL,
  ref_dir,
  layout = "layout",
  cores = 1
)
}
\arguments{
\item{doc_dir}{A directory containing PDFs}

\item{files}{A vector of PDF file names to be evaluated}

\item{ref_dir}{The name of the directory where reference extractions are to be exported}

\item{layout}{Specification of whether PDFs should be evaluated as no layout ("no_layout"), or layout with two columns ("layout")}

\item{cores}{how many cores you want to use in pblapply}
}
\value{
JSON files in ref_dir
}
\description{
Runs PDFs through the \href{https://anystyle.io/}{Anystyle} and create JSON files for each PDF with the identified references.
}
\examples{
reference_extract(doc_dir = 'documents_gsp/', ref_dir = 'reference_extracts_gsp', layout = "none")

}
