% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{working_references}
\alias{working_references}
\title{A sample of candidate references extracted from a PDF file that can be fed into reference_clean()}
\format{
A data.table object with 1001 rows
}
\description{
A dataset containing tagged references from a PDF extracted using anystyle.io.
}
