% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation_compile.R
\name{citation_compile}
\alias{citation_compile}
\title{Compiling JSONs from the Anystyle extraction}
\usage{
citation_compile(ref_dir)
}
\arguments{
\item{ref_dir}{directory name that holds the file(s) with .JSON extension extracted by Anystyle.io}
}
\value{
data frame
}
\description{
Reads in JSON files of extracted citations and combines them
}
\examples{
dt <- citation_compile('reference_extracts_gsp')


}
